RGB working and added reference bits.
