+++ keyboards/handwired/wylderbuilds/5x6_5_oled/config.h
@@ -28,11 +28,24 @@
 
 // WS2812 RGB LED strip input and number of LEDs
 #ifdef RGBLIGHT_ENABLE
-#define RGB_DI_PIN GP17
-#define RGBLED_NUM 34
+//#define WS2812_PIO_USE_PIO1
+//#define WS2812_PIO_USE_PIO1
+#define RGB_DI_PIN GP21
+//#define WS2812_EXTERNAL_PULLUP
+#define RGBLED_NUM 36
 //#define DRIVER_LED_TOTAL 34
-#define RGBLED_SPLIT { 17, 17 }
-#define STM32_SYSCLK KINETIS_SYSCLK_FREQUENCY
+#define RGBLED_SPLIT { 18, 18 }
+#define RGBLIGHT_SPLIT
+#define RGBLIGHT_EFFECT_RGB_TEST
+#define RGBLIGHT_EFFECT_BREATHING
+#define RGBLIGHT_EFFECT_RAINBOW_MOOD
+#define RGBLIGHT_EFFECT_SNAKE
+#define RGBLIGHT_EFFECT_TWINKLE
+#define RGBLIGHT_HUE_STEP 8
+#define RGBLIGHT_SAT_STEP 8
+#define RGBLIGHT_VAL_STEP 17
+#define RGBLIGHT_LIMIT_VAL 255
+//#define STM32_SYSCLK KINETIS_SYSCLK_FREQUENCY
 #define NOP_FUDGE 0.4
 #endif
 
@@ -41,6 +54,7 @@
 #define I2C_DRIVER I2CD1
 #define I2C1_SDA_PIN GP16
 #define I2C1_SCL_PIN GP17
+#define RGBLIGHT_LAYERS
 // OLED Options
 #define SPLIT_OLED_ENABLE
 #define SPLIT_WPM_ENABLE
