+++ keyboards/handwired/wylderbuilds/5x6_5_oled/rules.mk
@@ -2,9 +2,11 @@ MCU = RP2040
 SPLIT_KEYBOARD = yes
 BOOTLOADER = rp2040
 SERIAL_DRIVER = vendor
-RGBLIGHT_ENABLE = no
+RGBLIGHT_ENABLE = yes
+WS2812_DRIVER = vendor
+SLEEP_LED_ENABLE = no
 MOUSEKEY_ENABLE = yes
-CONSOLE_ENABLE = no
+CONSOLE_ENABLE = yes
 #COMMAND_ENABLE = yes
 BOOTMAGIC_ENABLE = yes
 OLED_ENABLE = yes
