+++ keyboards/handwired/wylderbuilds/reference/encoder/config.h
@@ -1,0 +1,8 @@
+// USE PINS 16 AND 17 TO USE 4 PIN JST, IF AVAILABLE
+
+#define ENCODERS_PAD_A { GP26 }
+#define ENCODERS_PAD_B { GP27 }
+#define ENCODER_RESOLUTIONS { 2 }
+//#define ENCODERS_PAD_A_RIGHT { GP26 }
+//#define ENCODERS_PAD_B_RIGHT { GP27 }
+//#define ENCODER_RESOLUTIONS_RIGHT { 2 }
