+++ keyboards/handwired/wylderbuilds/reference/encoder/rules.mk
@@ -1,0 +1,2 @@
+ENCODER_ENABLE = yes
+#ENCODER_MAP_ENABLE = yes
