+++ keyboards/handwired/wylderbuilds/reference/oled/config.h
@@ -1,0 +1,18 @@
+#ifdef OLED_ENABLE
+#define I2C_DRIVER I2CD1
+#define I2C1_SDA_PIN GP16
+#define I2C1_SCL_PIN GP17
+#define RGBLIGHT_LAYERS
+// OLED Options
+#define SPLIT_OLED_ENABLE
+#define SPLIT_WPM_ENABLE
+#define OLED_DISPLAY_WIDTH 128
+#define OLED_DISPLAY_HEIGHT 32
+#define OLED_MATRIX_SIZE 512
+#define OLED_RESET -1
+#define OLED_DISPLAY_ADDRESS 0x3C
+#define OLED_BRIGHTNESS 255
+#define OLED_TIMEOUT 32000
+#define OLED_FADE_OUT
+#define OLED_FADE_OUT_INTERVAL 0
+#endif
