+++ keyboards/handwired/wylderbuilds/reference/oled/rules.mk
@@ -1,0 +1,6 @@
+OLED_ENABLE = yes
+OLED_DRIVER = SSD1306
+WPM_ENABLE = yes
+EXTRAKEY_ENABLE = yes       # Audio control and System control
+
+OPT_DEFS += -DHAL_USE_I2C=TRUE
