+++ keyboards/handwired/wylderbuilds/reference/rgb_underglow/keymap.c
@@ -1,0 +1,54 @@
+
+// DEFINE THESE KEYS SOMEWHERE ON THE KEYMAP, USUALLY LEFT BOTTOM ROW
+//           _______,RGB_VAI,RGB_SAI,RGB_HUI,RGB_MOD,RGB_TOG,
+
+#ifdef RGBLIGHT_ENABLE
+const rgblight_segment_t PROGMEM querty_layer[] = RGBLIGHT_LAYER_SEGMENTS(
+    {15, 3, HSV_RED},       // Light 4 LEDs, starting with LED 6
+    {33, 3, HSV_RED}      // Light 4 LEDs, starting with LED 6
+);
+const rgblight_segment_t PROGMEM lower_layer[] = RGBLIGHT_LAYER_SEGMENTS(
+    {15, 3, HSV_BLUE},
+    {33, 3, HSV_BLUE}
+);
+const rgblight_segment_t PROGMEM raise_layer[] = RGBLIGHT_LAYER_SEGMENTS(
+    {15, 3, HSV_YELLOW},
+    {33, 3, HSV_YELLOW}
+);
+
+const rgblight_segment_t* const PROGMEM my_rgb_layers[] = RGBLIGHT_LAYERS_LIST(
+    querty_layer,    // Overrides caps lock layer
+    lower_layer,    // Overrides other layers
+    raise_layer     // Overrides other layers
+);
+
+//bool led_update_user(led_t led_state) {
+//    rgblight_set_layer_state(0, led_state.caps_lock);
+//    return true;
+//}
+
+layer_state_t default_layer_state_set_user(layer_state_t state) {
+    rgblight_set_layer_state(0, layer_state_cmp(state, _QWERTY));
+    return state;
+}
+
+layer_state_t layer_state_set_user(layer_state_t state) {
+    rgblight_set_layer_state(1, layer_state_cmp(state, _LOWER));
+    rgblight_set_layer_state(2, layer_state_cmp(state, _RAISE));
+    return state;
+}
+
+
+#endif
+
+//    rgblight_enable();
+#ifdef RGBLIGHT_ENABLE
+void keyboard_post_init_user(void) {
+    // Enable the LED layers
+    rgblight_layers = my_rgb_layers;
+}
+//    rgblight_layers = my_rgb_layers;
+//    rgblight_enable_noeeprom(); // Enables RGB, without saving settings
+//    rgblight_sethsv_noeeprom_cyan();
+//    rgblight_mode_noeeprom(RGBLIGHT_MODE_STATIC_LIGHT);
+#endif
