+++ keyboards/handwired/wylderbuilds/reference/rgb_underglow/rules.mk
@@ -1,0 +1,3 @@
+RGBLIGHT_ENABLE = yes
+WS2812_DRIVER = vendor
+SLEEP_LED_ENABLE = no
