+++ keyboards/handwired/wylderbuilds/trackball/5x7_track_underglow/config.h
@@ -17,7 +17,7 @@
 #ifdef RGBLIGHT_ENABLE
 //#define WS2812_PIO_USE_PIO1
 //#define WS2812_PIO_USE_PIO1
-#define RGB_DI_PIN GP28
+#define RGB_DI_PIN GP21
 //#define WS2812_EXTERNAL_PULLUP
 #define RGBLED_NUM 24
 //#define DRIVER_LED_TOTAL 34
@@ -43,7 +43,7 @@
 #define SERIAL_PIO_USE_PIO0
 #define SERIAL_USART_TIMEOUT     100  // USART driver timeout. default 100
 #define SERIAL_USART_SPEED 921600
-#define SERIAL_USART_PIN_SWAP
+//#define SERIAL_USART_PIN_SWAP
 
 #define RP2040_BOOTLOADER_DOUBLE_TAP_RESET
 #define RP2040_BOOTLOADER_DOUBLE_TAP_RESET_TIMEOUT 200U
@@ -63,16 +63,11 @@
 #define ROTATIONAL_TRANSFORM_ANGLE  -25
 #define POINTING_DEVICE_INVERT_Y
 #define POINTING_DEVICE_RIGHT
-#define PMW33XX_CS_PIN GP21
+#define PMW33XX_CS_PIN GP22
 #define POINTING_DEVICE_TASK_THROTTLE_MS 1
 #define PMW33XX_LIFTOFF_DISTANCE 0x05
 
 #define SPLIT_TRANSACTION_IDS_KB RPC_ID_KB_CONFIG_SYNC
-
-#define CHARYBDIS_MINIMUM_DEFAULT_DPI 1200
-#define CHARYBDIS_DEFAULT_DPI_CONFIG_STEP 200
-#define CHARYBDIS_MINIMUM_SNIPING_DPI 400
-#define CHARYBDIS_SNIPING_DPI_CONFIG_STEP 200
 
 
 //// Sensor Notes ////
